chore: update README for v3.0.1, remove console logs, and refactor event handling
