+++ scripts/apps/home/builder.js
@@ -1,21 +1,11 @@
 import { events, router } from "../../core/index.js";
 import { constants } from "../../utils/index.js";
+import { addBackhomeEvents } from "./events.js";
 
-function addBackhomeEvents(btn) {
-    events.registerEventListener(constants.ROUTES.hashHome, window.eventTracker, btn, 'click', () => {
-        if (window.location.hash === constants.ROUTES.hashHome) {
-            return;
-        }
-        router.changeRoute(constants.ROUTES.hashHome);
-    });
 
-}
 
 const backHomeBtn = (btnId) => {
-    const btn = document.getElementById(btnId);
-    addBackhomeEvents(btn);
-
-    return btn;
+    return addBackhomeEvents(document.getElementById(btnId));
 };
 export default {
     backHomeBtn,
